Game UI update
Game UI : GS-13
Create a rendrer Texture for the miniMap camera
Create a Material and add the texture render to the new material.
Create and position a new Camera to see the grid from the top.
Add the texture render to the target Texture of the miniMap Camera.
Create the MiniMap UI Frame and add thematerial into it to see the render of the miniMap Camera.