Kernel UI Show
Kernel UI : GS-15.
Add the kernel UI to the UI Links.
Create method for Hide, Show the Kernel UI.
Add Action of Showing the Kernel UI to the click action on a Kernel object.
Add action of hiding the Kernel UI to click action on another element of the grid.